data_analyst_multi_agent/streamlit/requirements.txt
streamlit 
agno 
google-genai 
duckdb 
pandas 
matplotlib 
pyarrow 
python-dotenv
